feat: create Carousel component
